r2.11 cherry-pick: 6965f231b8b "Replace ptrdiff_t with int64_t."

+++ tensorflow/core/lib/gif/gif_io_test.cc
@@ -52,7 +52,7 @@ void TestDecodeGif(Env* env, DecodeGifTestCase testcase) {
         w = width;
         h = height;
         c = channels;
-        return new uint8[ptrdiff_t(frame_cnt) * height * width * channels];
+        return new uint8[int64_t(frame_cnt) * height * width * channels];
       },
       &error_string));
   ASSERT_NE(imgdata, nullptr);
